Read AWS credentials from environment variables.
If the credentials are not set in the configuration file, bifrost will
look for credentials in the AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY
environment variables.

These are the same environment variables used by the AWS SDKs as of July 2014.
(http://blogs.aws.amazon.com/security/post/Tx3D6U6WSFGOK2H/A-New-and-Standardized-Way-to-Manage-Credentials-in-the-AWS-SDKs)