Protect from failures in bucket creation.
The scenario this mitigates is when we start a number of bifrost
instances simultaneously and they race to create the bucket. The losers
get an exception and then fail to start the upload thread, meaning they
never process anything.

We could implement a retry here, but it seems over complicated for this
use case. With this change the Worse case is that we get an exception
later that the bucket doesn't exist.